InteropBroker refactor
Broken out some of the helper functions from the interopbroker and also removed the intent related functions from the app provider.
