+++ how-to/workspace-platform-starter/CHANGELOG.md
@@ -41,7 +41,9 @@
 - Breaking Change - We have updated the way an interop broker communicates with an intent picker window. We were using a reference to the WebWindow to add a function from the broker to the intent picker window. This works but would not work cross domain. OpenFin also introduced the showPopupWindow API since this approach was used so we have now switched to using the showPopupWindow which works cross domain. This means that you can not build you own intent window that receives the list of applications and intent/context via customData passed to the window and it can return the selection using await fin.me.dispatchPopupResult({ appId, instanceId, intent }); if a selection is made or await fin.me.dispatchPopupResult(); if the user cancelled the intent selection. Our example intent pickers have been updated: [instance-picker.html](./public/common/windows/intents/instance-picker.html) and [picker.html](./public/common/windows/intents/picker.html)
 - New Feature - The intent picker now tries to appear in the center of the monitor from where the intent was raised (rather than always showing up on the main monitor).
 - Fix - Fixed the basic intent picker that wasn't working correctly.
-- Added Tray menu can now use the custom popup menus, will use the global popupMenuStyle but can be overriden by the `trayProvider.popupMenuStyle`
+- Added Tray menu can now use the custom popup menus, will use the global popupMenuStyle but can be overridden by the `trayProvider.popupMenuStyle`
+- Updated interopbroker to break out some logic into sub classes in a broker folder. Also removed functions that were not being overridden.
+- Removed interop related functions from [apps.ts](./client/src/framework/apps.ts) to [app-intent-helper.ts](./client/src/framework/platform/broker/app-intent-helper.ts) as part of the interopbroker tidy.
 
 ## v14
