+++ how-to/workspace-platform-starter/client/types/module/shapes/interopbroker-shapes.d.ts
@@ -85,9 +85,9 @@ export type IntentTargetMetaData =
 			instanceId?: string;
 	  };
 /**
- * The response from the intent picker.
+ * The response from the intent resolver.
  */
-export interface IntentPickerResponse {
+export interface IntentResolverResponse {
 	/**
 	 * The appId that was selected.
 	 */
